requirements.txt
blinker==1.4
bson==0.5.10
certifi==2020.4.5.2
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Admin==1.5.6
Flask-Cache==0.13.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-OAuthlib==0.9.5
Flask-PyMongo==2.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
flask-swagger==0.2.14
flask-swagger-ui==3.25.0
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.9
itsdangerous==0.24
jieba==0.42.1
Jinja2==2.9.6
MarkupSafe==1.0
mysql-connector-python==8.0.20
mysqlclient==1.4.6
numpy==1.18.5
oauthlib==2.1.0
protobuf==3.12.2
pymongo==3.10.1
python-dateutil==2.8.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==0.12.2
Whoosh==2.7.4
wincertstore==0.2
WTForms==2.1
